requirements.txt
# Core Flask
Flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0

# Database
SQLAlchemy>=2.0.0

# Authentication & Security
bcrypt>=4.0.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

# HTTP & WebSockets
requests>=2.31.0
websocket-client>=1.6.0

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

